+++ src/main.rs
@@ -2,13 +2,25 @@ mod acquire;
 mod plot;
 
 use argh::FromArgs;
+use csv::Writer;
+use once_cell::sync::Lazy;
 use std::io::{stdout, Read, StdoutLock, Write};
+use std::time::SystemTime;
 use std::{error::Error, thread::sleep, time::Duration};
 use termion::{
     async_stdin, clear, cursor,
     cursor::Goto,
     raw::{IntoRawMode, RawTerminal},
 };
+
+/// Determine duration since epoch, once, lazily.
+/// When `Lazy` is stabilized in `std::sync::Lazy` this can move over to that.
+pub static TIMESTAMP: Lazy<String> = Lazy::new(|| {
+    let secs_epoch = SystemTime::now()
+        .duration_since(SystemTime::UNIX_EPOCH)
+        .expect("failed to get duration since epoch");
+    secs_epoch.as_secs().to_string()
+});
 
 #[derive(FromArgs)]
 /// Simple tool to sample and plot power consumption, average frequency and cpu die temperatures over time.
@@ -17,16 +29,20 @@ pub struct Arghs {
     #[argh(option, short = 'i', default = "1000")]
     interval: u64,
 
+    /// optionally save data series (defaults to `true`)
+    #[argh(option, short = 'w', default = "true")]
+    write_data: bool,
+
     /// optional title (e.g. a condition for the run)
     #[argh(
         option,
         short = 't',
-        default = "String::from(\"thermals and performance under load\")"
+        default = "String::from(\"thermals / avg. core clocks under load\")"
     )]
     title: String,
 
     /// optional image size dimensions WxH (1024x768)
-    #[argh(option, from_str_fn(into_plot_dimensions), default = "(1024, 768)")]
+    #[argh(option, from_str_fn(into_plot_dimensions), default = "(1280, 768)")]
     wxh: (u32, u32),
 }
 
@@ -88,8 +104,28 @@ fn main() -> Result<(), Box<dyn Error>> {
 
     write!(stdout, "{}{}Saving plots..", Goto(1, 8), clear::AfterCursor)?;
     plot::plot(&args, &freq_series, &pwr_series, &temp_series)?;
-    write!(stdout, " done!{}", Goto(1, 12))?;
 
+    if args.write_data {
+        write!(stdout, "{}{}Saving csv..", Goto(1, 10), clear::AfterCursor)?;
+        let filename_csv = format!("data_series_{}.csv", *TIMESTAMP);
+        let mut wtr = Writer::from_path(&filename_csv)?;
+
+        // Write column names
+        wtr.serialize(&["Power", "Clock", "Temperature"])?;
+
+        //  A sample of each quantity together makes a sample record.
+        // This loops iterates over these samples and writes them to the CSV.
+        for sample in pwr_series
+            .iter()
+            .zip(freq_series.iter())
+            .zip(temp_series.iter())
+        {
+            wtr.serialize(sample)?;
+        }
+        wtr.flush()?;
+    }
+
+    write!(stdout, " done!{}", Goto(1, 12))?;
     stdout.flush()?;
     Ok(())
 }
@@ -100,8 +136,8 @@ fn print_values(
     power: f64,
     freq: f64,
 ) -> std::io::Result<()> {
-    write!(stdout, "{}Avg. CPU frequency: {:0.2} MHz", Goto(1, 5), freq)?;
-    write!(stdout, "{}CPU die temp: {:0.2} °C", Goto(1, 3), temp)?;
-    write!(stdout, "{}CPU power: {: >#4.2} Watt", Goto(1, 4), power)?;
+    write!(stdout, "{}Avg. CPU frequency: {:0.2} MHz", Goto(1, 3), freq)?;
+    write!(stdout, "{}CPU die temp: {:0.2} °C", Goto(1, 4), temp)?;
+    write!(stdout, "{}CPU power: {: >#4.2} Watt", Goto(1, 5), power)?;
     Ok(())
 }