+++ CHANGELOG.md
@@ -1,7 +1,20 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.1]
+
+* New feature
+  * save collected data as `cvs`
+
+* Plot
+  * somewhat better default title
+  * larger default plot size
+  * thicker strokes
+  * horizontal mesh
+
+* Internal
+  * Lazily determine time since epoch